refactor(button): move variants to separate file for better organization
