+++ apps/demo/src/components/starwind/button/Button.astro
@@ -1,7 +1,7 @@
 ---
 import type { HTMLAttributes } from "astro/types";
 import { type VariantProps } from "tailwind-variants";
-import { buttonVariants } from "./index";
+import { buttonVariants } from "./variants";
 
 interface Props
   extends HTMLAttributes<"button">,
